Algorith for Rpeak detection done
Created recentRR and recentRR_OK buffers!

+++ ECG-Scanner/bufferPeak.h
@@ -3,9 +3,9 @@
 #include "peak.h"
 
 typedef struct{
-	int head;
-	int size;
-	int inserts;
+	unsigned int head;
+	unsigned int size;
+	unsigned long int inserts;
 	Peak * data;
 } buffPeak;
